Fading between countdown video and applause video
Fading and GPIO control of countdown video has been moved to the
player method "fade", which causes the fading and GPIO actions to
be independent from the state machine's states.
But the moment for selecting and initialising the applause video
should be even earlier to get an even better behaviour.