Bugfix when pressing the buzzer at the wrong time
When pressing the buzzer while the video is fading from countdown to
applause there occureed an error which resulted in hanging video
playback. The reason was that the countdown video was still active
and playing when the new buzzer impulse was detected.
It was fixed by enabling the buzzer AFTER the countdown video
was completely unloaded.

open issue:
- config-/meta files for fading parameters
  either locally for each video file
  or globally if no local file exists

System requirements:
This software doesn't run satisfying on a RPi0